security + some optimization

+++ src/main/java/pl/edziennik/edziennik/mark/MarkController.java
@@ -10,6 +10,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import pl.edziennik.edziennik.mark.category.MarkCategory;
 import pl.edziennik.edziennik.mark.category.MarkCategoryRepository;
+import pl.edziennik.edziennik.security.LoggedUser;
 import pl.edziennik.edziennik.student.Student;
 import pl.edziennik.edziennik.student.StudentRepository;
 import pl.edziennik.edziennik.teacher.TeacherRepository;
@@ -20,8 +21,6 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 @Controller
@@ -31,6 +30,7 @@ public class MarkController {
     private final StudentRepository studentRepository;
     private final MarkCategoryRepository markCategoryRepository;
     private final TeacherRepository teacherRepository;
+    private final LoggedUser loggedUser;
     @Autowired
     private final EntityManager entityManager;
 
@@ -38,16 +38,21 @@ public class MarkController {
                           StudentRepository studentRepository,
                           MarkCategoryRepository markCategoryRepository,
                           TeacherRepository teacherRepository,
+                          LoggedUser loggedUser,
                           EntityManager entityManager) {
         this.markRepository = markRepository;
         this.studentRepository = studentRepository;
         this.markCategoryRepository = markCategoryRepository;
         this.teacherRepository = teacherRepository;
+        this.loggedUser = loggedUser;
         this.entityManager = entityManager;
     }
 
     @GetMapping("/add/{mark}/{studentId}/{markCategoryId}/{markId}")
     public ResponseEntity<Boolean> add(@PathVariable String mark, @PathVariable Long studentId, @PathVariable Long markCategoryId, @PathVariable String markId) {
+        if (!loggedUser.teachesSubjectOfMarkCategoryId(markCategoryId)) {
+            return null;
+        }
         Mark markObj = new Mark();
         Student student = studentRepository.getReferenceById(studentId);
         Mark probableMark = student.getMarkByMarkCategoryId(markCategoryId);
@@ -70,22 +75,26 @@ public class MarkController {
 
     @GetMapping("/{id}/history")
     public String idHistory(@PathVariable Long id) {
+        if (!loggedUser.hasAccessToMarkHistory(id)) {
+            return "error/403";
+        }
         Mark mark = markRepository.getReferenceById(id);
         return "redirect:/mark/history/" + mark.getMarkCategory().getId() + "/" + mark.getStudent().getId();
     }
 
     @GetMapping("/history/{categoryId}/{studentId}")
     public String history(@PathVariable Long categoryId, @PathVariable Long studentId, Model model) {
+        Mark mark = markRepository.getMarkByStudentIdAndMarkCategoryId(studentId, categoryId);
+        if (!loggedUser.hasAccessToMarkHistory(mark.getId())) {
+            return "error/403";
+        }
         Student student = studentRepository.getReferenceById(studentId);
         MarkCategory markCategory = markCategoryRepository.getReferenceById(categoryId);
         AuditReader reader = AuditReaderFactory.get(entityManager);
         List<Mark> history = new ArrayList<>();
-        Mark mark = markRepository.getMarkByStudentIdAndMarkCategoryId(studentId, categoryId);
-        if (mark != null) {
-            List<Number> revisions = reader.getRevisions(mark.getClass(), mark.getId());
-            for (Number revision : revisions) {
-                history.add(reader.find(Mark.class, mark.getId(), revision));
-            }
+        List<Number> revisions = reader.getRevisions(mark.getClass(), mark.getId());
+        for (Number revision : revisions) {
+            history.add(reader.find(Mark.class, mark.getId(), revision));
         }
         model.addAttribute("history", history);
         model.addAttribute("student", student);
@@ -98,6 +107,9 @@ public class MarkController {
     public ResponseEntity<Boolean> delete(@PathVariable Long categoryId, @PathVariable Long studentId) {
         Student student = studentRepository.getReferenceById(studentId);
         Mark mark = student.getMarkByMarkCategoryId(categoryId);
+        if (!loggedUser.isMarkCreator(mark.getId())) {
+            return null;
+        }
         markRepository.delete(mark);
         return new ResponseEntity<>(true, HttpStatus.OK);
     }
